Updated files.
The logic for array expansion is now contained in a private method, expandArray().
Wrote the push() method to add an element to the queue and resize it if it's full.
Both of these methods must be tested because it's quite possible I'm not understanding
something correctly, or some other error is occurring.